+++ Shape_detection/include/CGAL/Shape_detection/Region_growing/Region_growing_on_face_graph/Propagation_conditions_on_face_graph.h
@@ -74,7 +74,7 @@ namespace CGAL {
             using Get_sqrt               = CGAL::Shape_detection::Get_sqrt<Traits>;
             using Sqrt                   = typename Get_sqrt::Sqrt;
 
-            using Index                  = int;
+            using Index                  = long;
             ///< \endcond
 
             /*!
